refactor: recompose and memoise
