+++ apps/client/src/features/cuesheet/cuesheet-table-elements/CuesheetRow.tsx
@@ -1,0 +1,6 @@
+interface CuesheetRowProps {
+  row: OntimeRundownEntry;
+  isSelected: boolean;
+}
+
+function CuesheetRow() {}
