+++ apps/client/src/features/cuesheet/cuesheet-table-elements/EditableCell.tsx
